requirements.txt
Django>=4.2
celery>=5.5
redis>=5.0
requests>=2.31
pdfplumber>=0.8.0
pytesseract>=0.3.10
PyMuPDF>=1.22.0
Pillow>=10.0
python-dotenv>=1.0
supabase>=2.0.0
djangorestframework>=3.15.0
PyJWT>=2.8.0
dj-database-url>=1.0.0
